+++ R/MAINest.R
@@ -5,9 +5,11 @@
 #'   phase. Parallel computing is used to perform multiple rounds of estimations in parallel.
 #'
 #' @inheritParams GAfit
-#' @param ncalls number of estimation rounds that should be performed.
-#' @param ncores number cores to be used in parallel computing.
-#' @param maxit maximum number of iterations in the variable metric algorithm.
+#' @param ncalls the number of estimation rounds that should be performed.
+#' @param ncores the number cores to be used in parallel computing.
+#' @param maxit the maximum number of iterations in the variable metric algorithm.
+#' @param seeds a length \code{ncalls} vector containing the random number generator seed for each call to the genetic algorithm,
+#'   or \code{NULL} for not initializing the seed. Exists for creating reproducible results.
 #' @param print_res should summaries of estimation results be printed?
 #' @param ... additional settings passed to the function \code{GAfit} employing the genetic algorithm.
 #' @details
@@ -94,8 +96,9 @@
 #' data <- cbind(10*eurusd[,1], 100*eurusd[,2])
 #' colnames(data) <- colnames(eurusd)
 #'
-#' # GMVAR(1,2) model with default settings
-#' fit12 <- fitGMVAR(data, p=1, M=2)
+#' # GMVAR(1,2) model: 10 estimation rounds with seeds set
+#' # for reproducibility
+#' fit12 <- fitGMVAR(data, p=1, M=2, ncalls=10, seeds=1:10)
 #' fit12
 #' plot(fit12)
 #' summary(fit12)
@@ -124,9 +127,10 @@
 
 
 fitGMVAR <- function(data, p, M, conditional=TRUE, parametrization=c("intercept", "mean"), constraints=NULL, ncalls=round(10 + 9*log(M)),
-                     ncores=min(ncalls, parallel::detectCores()), maxit=300, print_res=TRUE, ...) {
+                     ncores=min(2, ncalls, parallel::detectCores()), maxit=300, seeds=NULL, print_res=TRUE, ...) {
 
   on.exit(closeAllConnections())
+  if(!is.null(seeds) && length(seeds) != ncalls) stop("The argument 'seeds' needs be NULL or a vector of length 'ncalls'")
   parametrization <- match.arg(parametrization)
   if(!all_pos_ints(c(p, M, ncalls, ncores, maxit))) stop("Arguments p, M, ncalls, ncores and maxit must be positive integers")
   data <- check_data(data=data, p=p)
@@ -154,15 +158,15 @@ fitGMVAR <- function(data, p, M, conditional=TRUE, parametrization=c("intercept"
   parallel::clusterEvalQ(cl, c(library(Brobdingnag), library(mvnfast), library(pbapply)))
 
   cat("Optimizing with genetic algorithm...", "\n")
-  GAresults <- pbapply::pblapply(1:ncalls, function(x) GAfit(data=data, p=p, M=M, conditional=conditional, parametrization=parametrization,
-                                                              constraints=constraints, ...), cl=cl)
+  GAresults <- pbapply::pblapply(1:ncalls, function(i1) GAfit(data=data, p=p, M=M, conditional=conditional, parametrization=parametrization,
+                                                              constraints=constraints, seed=seeds[i1], ...), cl=cl)
   parallel::stopCluster(cl=cl)
 
   loks <- vapply(1:ncalls, function(i1) loglikelihood_int(data, p, M, params=GAresults[[i1]], conditional=conditional,
                                                           parametrization=parametrization, constraints=constraints,
                                                           check_params=TRUE, to_return="loglik", minval=minval), numeric(1))
 
-  if(print_res == TRUE) {
+  if(print_res) {
     cat("Results from genetic algorithm:", "\n")
     cat(paste("lowest value: ", round(min(loks), 3)), "\n")
     cat(paste("mean value:   ", round(mean(loks), 3)), "\n")
@@ -196,7 +200,7 @@ fitGMVAR <- function(data, p, M, conditional=TRUE, parametrization=c("intercept"
   loks <- vapply(1:ncalls, function(i1) loglikelihood_int(data=data, p=p, M=M, params=NEWTONresults[[i1]]$par, conditional=conditional,
                                                           constraints=constraints, parametrization=parametrization, check_params=TRUE,
                                                           to_return="loglik", minval=minval), numeric(1))
-  if(print_res == TRUE) {
+  if(print_res) {
     cat("Results from variable metric algorithm:\n")
     cat(paste("lowest value: ", round(min(loks), 3)), "\n")
     cat(paste("mean value:   ", round(mean(loks), 3)), "\n")