+++ NEWS.md
@@ -8,3 +8,10 @@
 * Generally more functionality for conditional and unconditional moments.
 * Increased numerical tolerance for considering models to be stationary to attain better numerical stability.
 * Updates on documentation, fixed typos.
+
+# gmvarkit 1.1.1
+
+* Changed the default number of CPU cores employed by the estimation function fitGMVAR to be at most two due to CRAN policy.
+* Added argument "seeds" to fitGMVAR allowing one to set the random number generator seed for each call to the genetic algorithm.
+* New exported function: alt_gmvar
+* Updated the vignette to include the definition of the GMVAR model